Crate rome_formatter

Infrastructure for code formatting

This module defines FormatElement, an IR to format code documents and provides a mean to print such a document to a string. Objects that know how to format themselves implement the Format trait.

Formatting Traits

Format: Implemented by objects that can be formatted.

FormatRule: Rule that knows how to format an object of another type. Necessary in the situation where it’s necessary to implement Format on an object from another crate. This module defines the FormatRefWithRule and FormatOwnedWithRule structs to pass an item with its corresponding rule.

FormatWithRule implemented by objects that know how to format another type. Useful for implementing some reusable formatting logic inside of this module if the type itself doesn’t implement Format

Formatting Macros

This crate defines two macros to construct the IR. These are inspired by Rust’s fmt macros

format!: Formats a formatable object

format_args!: Concatenates a sequence of Format objects.

write!: Writes a sequence of formatable objects into an output buffer.

Re-exports

pub use format_element::normalize_newlines;

pub use format_element::FormatElement;

pub use format_element::Token;

pub use format_element::Verbatim;

pub use format_element::LINE_TERMINATORS;

pub use group_id::GroupId;

Modules

format_element

formatter

group_id

intersperse

Copied from Rust’s unstable iter.intersperse().

macros

prelude

printed_tokens

printer

token

Macros

best_fitting

Provides multiple different alternatives and the printer picks the first one that fits. Use this as last resort because it requires that the printer must try all variants in the worst case. The passed variants must be in the following order:

dbg_write

Writes formatted data into the given buffer and prints all written elements for a quick and dirty debugging.

format

Creates the Format IR for a value.

format_args

Constructs the parameters for other formatting macros.

write

Writes formatted data into a buffer.

Structs

Argument

Mono-morphed type to format an object. Used by the crate::format!, crate::format_args!, and crate::write! macros.

Arguments

Sequence of objects that should be formatted in the specified order.

BestFitting

The first variant is the most flat, and the last is the most expanded variant. See best_fitting! macro for a more in-detail documentation

FormatOwnedWithRule

Formats the item with the specified rule.

FormatRefWithRule

Formats the referenced item with the specified rule.

FormatState

This structure stores the state that is relevant for the formatting of the whole document.

FormatStateSnapshot

Formatted

HasLabelBuffer

Inspect

Buffer that allows you inspecting elements as they get written to the formatter.

LastTokenKind

LineWidth

Validated value for the line_width formatter options

LineWidthFromIntError

Error type returned when converting a u16 to a LineWidth fails

PreambleBuffer

This struct wraps an existing buffer and emits a preamble text when the first text is written.

Printed

SimpleFormatContext

SourceComment

SourceMarker

Lightweight sourcemap marker between source and output tokens

VecBuffer

Vector backed Buffer implementation.

WillBreakBuffer

Enums

BufferSnapshot

Snapshot of a buffer state that can be restored at a later point.

CommentKind

FormatError

Series of errors encountered during formatting

IndentStyle

ParseLineWidthError

Error type returned when parsing a LineWidth from a string fails

Traits

Buffer

A trait for writing or formatting into FormatElement-accepting buffers or streams.

BufferExtensions

CommentContext

Stores comments specific context information.

CommentStyle

Defines how to format comments for a specific [Language].

Format

Formatting trait for types that can create a formatted representation. The rome_formatter equivalent to std::fmt::Display.

FormatContext

Context configuring how an object gets formatted.

FormatRule

Rule that knows how to format an object of type T.

FormatRuleWithOptions

Rule that supports customizing how it formats an object of type T.

FormatWithRule

Trait for an object that formats an object with a specified rule.

Functions

block_indent

Inserts a hard line break before and after the content and increases the indention level for the content by one.

comment

Marks some content as a comment trivia.

empty_line

A forced empty line. An empty line inserts enough line breaks in the output for the previous and next element to be separated by an empty line.

format

The format function takes an Arguments struct and returns the resulting formatting IR.

format_node

Formats a syntax node file based on its features.

format_range

Formats a range within a file, supported by Rome

format_sub_tree

Formats a single node within a file, supported by Rome.

get_lines_before

Get the number of line breaks between two consecutive SyntaxNodes in the tree

group_elements

Creates a logical Group around the content that should either consistently be printed on a single line or broken across multiple lines.

hard_line_break

A forced line break that are always printed. A hard line break forces any enclosing Group to be printed over multiple lines.

if_group_breaks

Adds a conditional content that is emitted only if it isn’t inside an enclosing Group that is printed on a single line. The element allows, for example, to insert a trailing comma after the last array element only if the array doesn’t fit on a single line.

if_group_fits_on_line

Adds a conditional content specific for Groups that fit on a single line. The content isn’t emitted for Groups spanning multiple lines.

indent

It adds a level of indentation to the given content

labelled

Marks some content with a label.

line_suffix

Pushes some content to the end of the current line

soft_block_indent

Indents the content by inserting a line break before and after the content and increasing the indention level for the content by one if the enclosing group doesn’t fit on a single line. Doesn’t change the formatting if the enclosing group fits on a single line.

soft_line_break

A line break that only gets printed if the enclosing Group doesn’t fit on a single line. It’s omitted if the enclosing Group fits on a single line. A soft line break is identical to a hard line break when not enclosed inside of a Group.

soft_line_break_or_space

A line break if the enclosing Group doesn’t fit on a single line, a space otherwise.

soft_line_indent_or_space

If the enclosing Group doesn’t fit on a single line, inserts a line break and indent. Otherwise, just inserts a space.

space_token

Inserts a single space. Allows to separate different tokens.

token

Creates a token that gets written as is to the output. Make sure to properly escape the text if it’s user generated (e.g. a string and not a language keyword).

write

The write function takes a target buffer and an Arguments struct that can be precompiled with the format_args! macro.

Type Definitions

FormatResult

Public return type of the formatter