Enum rome_formatter::comments::CommentPlacement

pub enum CommentPlacement<L: Language> {
    Leading {
        node: SyntaxNode<L>,
        comment: SourceComment<L>,
    },
    Trailing {
        node: SyntaxNode<L>,
        comment: SourceComment<L>,
    },
    Dangling {
        node: SyntaxNode<L>,
        comment: SourceComment<L>,
    },
    Default(DecoratedComment<L>),
}

Variants

Leading

Fields

node: SyntaxNode<L>

comment: SourceComment<L>

Makes comment a leading comment of node.

Trailing

Fields

node: SyntaxNode<L>

comment: SourceComment<L>

Makes comment a trailing comment of node.

Dangling

Fields

node: SyntaxNode<L>

comment: SourceComment<L>

Makes comment a dangling comment of node.

Default(DecoratedComment<L>)

Uses the default heuristic to determine the placement of the comment.

Same line comments

Makes the comment a…

trailing comment of the preceding_node if both the following_node and preceding_node are not None and the comment and preceding_node are only separated by a space (there’s no token between the comment and preceding_node).

leading comment of the following_node if the following_node is not None

trailing comment of the preceding_node if the preceding_node is not None

dangling comment of the enclosing_node.

Examples

Comment with preceding and following nodes

[
    a, // comment
    b
]

The comment becomes a trailing comment of the node a.

Comment with preceding node only

[
    a // comment
]

The comment becomes a trailing comment of the node a.

Comment with following node only

[ // comment
    b
]

The comment becomes a leading comment of the node b.

Dangling comment

[ // comment
]

The comment becomes a dangling comment of the enclosing array expression because both the preceding_node and following_node are None.

Own line comments

Makes the comment a…

leading comment of the following_node if the following_node is not None

or a trailing comment of the preceding_node if the preceding_node is not None

or a dangling comment of the enclosing_node.

Examples

Comment with leading and preceding nodes

[
    a,
    // comment
    b
]

The comment becomes a leading comment of the node b.

Comment with preceding node only

[
    a
    // comment
]

The comment becomes a trailing comment of the node a.

Comment with following node only

[
    // comment
    b
]

The comment becomes a leading comment of the node b.

Dangling comment

[
    // comment
]

The comment becomes a dangling comment of the array expression because both preceding_node and following_node are None.

End of line comments

Makes the comment a…

trailing comment of the preceding_node if the preceding_node is not None

or a leading comment of the following_node if the following_node is not None

or a dangling comment of the enclosing_node.

Examples

Comment with leading and preceding nodes

[a /* comment */, b]

The comment becomes a trailing comment of the node a because there’s no token between the node a and the comment.

[a, /* comment */ b]

The comment becomes a leading comment of the node b because the node a and the comment are separated by a , token.

Comment with preceding node only

[a, /* last */ ]

The comment becomes a trailing comment of the node a because the following_node is None.

Comment with following node only

[/* comment */ b]

The comment becomes a leading comment of the node b because the preceding_node is None

Dangling comment

[/* comment*/]

The comment becomes a dangling comment of the array expression because both preceding_node and following_node are None.

Implementations

impl<L: Language> CommentPlacement<L>

pub fn leading(node: SyntaxNode<L>, comment: impl Into<SourceComment<L>>) -> Self

Makes comment a leading comment of node.

pub fn dangling(
node: SyntaxNode<L>,
comment: impl Into<SourceComment<L>>
) -> Self

Makes comment a dangling comment of node.

pub fn trailing(
node: SyntaxNode<L>,
comment: impl Into<SourceComment<L>>
) -> Self

Makes comment a trailing comment of node.

pub fn or_else<F>(self, f: F) -> Self where
F: FnOnce(DecoratedComment<L>) -> CommentPlacement<L>,

Returns the placement if it isn’t CommentPlacement::Default, otherwise calls f and returns the result.