Struct rome_formatter::comments::Comments

pub struct Comments<L: Language> { /* private fields */ }

The comments of a syntax tree stored by node.

Cloning comments is cheap as it only involves bumping a reference counter.

Implementations

impl<L: Language> Comments<L>

pub fn from_node<Style>(
root: &SyntaxNode<L>,
style: &Style,
source_map: Option<&TransformSourceMap>
) -> Self where
Style: CommentStyle<Language = L>,

Extracts all the comments from root and its descendants nodes.

pub fn has_comments(&self, node: &SyntaxNode<L>) -> bool

Returns true if the given node has any leading or trailing comments.

pub fn has_leading_comments(&self, node: &SyntaxNode<L>) -> bool

Returns true if the given node has any leading comments.

pub fn has_leading_own_line_comment(&self, node: &SyntaxNode<L>) -> bool

Tests if the node has any leading comments that have a leading line break.

Corresponds to CommentTextPosition::OwnLine.

pub fn leading_comments(&self, node: &SyntaxNode<L>) -> &[SourceComment<L>]

Returns the node’s leading comments.

pub fn has_dangling_comments(&self, node: &SyntaxNode<L>) -> bool

Returns true if node has any dangling comments.

pub fn dangling_comments(&self, node: &SyntaxNode<L>) -> &[SourceComment<L>]

Returns the dangling comments of node

pub fn trailing_comments(&self, node: &SyntaxNode<L>) -> &[SourceComment<L>]

Returns the node’s trailing comments.

pub fn has_trailing_line_comment(&self, node: &SyntaxNode<L>) -> bool

Returns true if the node has any trailing line comment.

pub fn has_trailing_comments(&self, node: &SyntaxNode<L>) -> bool

Returns true if the given node has any trailing comments.

pub fn leading_trailing_comments(
&self,
node: &SyntaxNode<L>
) -> impl Iterator<Item = &SourceComment<L>>

Returns an iterator over the leading and trailing comments of node.

pub fn leading_dangling_trailing_comments<'a>(
&'a self,
node: &'a SyntaxNode<L>
) -> impl Iterator<Item = &'_ SourceComment<L>> + 'a

Returns an iterator over the leading, dangling, and trailing comments of node.

pub fn has_skipped(&self, token: &SyntaxToken<L>) -> bool

Returns true if that node has skipped token trivia attached.

pub fn is_suppressed(&self, node: &SyntaxNode<L>) -> bool

Returns true if node has a leading, dangling, or trailing suppression comment.

Examples

// rome-ignore format: Reason
console.log("Test");

Returns true for the expression statement but false for the call expression because the call expression is nested inside of the expression statement.

pub fn mark_suppression_checked(&self, node: &SyntaxNode<L>)

Marks that it isn’t necessary for the given node to check if it has been suppressed or not.