Struct rome_formatter::comments::SourceComment

pub struct SourceComment<L: Language> { /* private fields */ }

A comment in the source document.

Implementations

impl<L: Language> SourceComment<L>

pub fn piece(&self) -> &SyntaxTriviaPieceComments<L>

Returns the underlining comment trivia piece

pub fn lines_before(&self) -> u32

The number of lines between this comment and the previous token or comment.

Examples

Same line

a // end of line

Returns 0 because there’s no line break between the token a and the comment.

Own Line

a;

/* comment */

Returns 2 because there are two line breaks between the token a and the comment.

pub fn lines_after(&self) -> u32

The number of line breaks right after this comment.

Examples

End of line

a; // comment

b;

Returns 2 because there are two line breaks between the comment and the token b.

Same line

a;
/* comment */ b;

Returns 0 because there are no line breaks between the comment and the token b.

pub fn kind(&self) -> CommentKind

The kind of the comment

pub fn mark_formatted(&self)

Marks the comment as formatted