Trait rome_formatter::comments::CommentStyle

pub trait CommentStyle: Default {
    type Language: Language;

    fn get_comment_kind(
        comment: &SyntaxTriviaPieceComments<Self::Language>
    ) -> CommentKind;

    fn is_suppression(_text: &str) -> bool { ... }
    fn place_comment(
        &self,
        comment: DecoratedComment<Self::Language>
    ) -> CommentPlacement<Self::Language> { ... }
}

Defines how to format comments for a specific Language.

Required Associated Types

type Language: Language

Required Methods

fn get_comment_kind(
comment: &SyntaxTriviaPieceComments<Self::Language>
) -> CommentKind

Returns the (kind)CommentKind of the comment

Provided Methods

fn is_suppression(_text: &str) -> bool

Returns true if a comment with the given text is a rome-ignore format: suppression comment.

fn place_comment(
&self,
comment: DecoratedComment<Self::Language>
) -> CommentPlacement<Self::Language>

Determines the placement of comment.

The default implementation returns CommentPlacement::Default.