Function rome_formatter::format_range

pub fn format_range<Context, L, R, P>(
    context: Context, 
    root: &SyntaxNode<L>, 
    range: TextRange, 
    predicate: P
) -> FormatResult<Printed> where
    Context: FormatContext,
    L: Language,
    R: FormatRule<SyntaxNode<L>, Context = Context> + Default,
    P: FnMut(&SyntaxNode<L>) -> bool, 

Formats a range within a file, supported by Rome

Because this function is language-agnostic, it must be provided with a predicate function that’s used to select appropriate “root nodes” for the range formatting process: for instance in JavaScript the predicate returns true for statement and declaration nodes, to ensure the entire statement gets formatted instead of the smallest sub-expression that fits the range

This runs a simple heuristic to determine the initial indentation level of the node based on the provided FormatContext, which must match currently the current initial of the file. Additionally, because the reformatting happens only locally the resulting code will be indented with the same level as the original selection, even if it’s a mismatch from the rest of the block the selection is in

It returns a Formatted result with a range corresponding to the range of the input that was effectively overwritten by the formatter