Module rome_formatter::format_element

Structs

BestFitting

Provides the printer with different representations for the same element so that the printer can pick the best fitting variant.

ConditionalGroupContent

Fill

Fill is a list of FormatElements along with a separator.

Group

Group is a special token that controls how the child tokens are printed.

Interned

Label

LabelId

List

A token used to gather a list of elements; see crate::Formatter::join_with.

Verbatim

Information of the node/token formatted verbatim

Enums

FormatElement

Language agnostic IR for formatting source code.

LineMode

PrintMode

Token

See crate::builders::token for documentation

VerbatimKind

Constants

LINE_TERMINATORS

Functions

normalize_newlines

Replace the line terminators matching the provided list with “\n” since its the only line break type supported by the printer