Enum rome_formatter::format_element::FormatElement

pub enum FormatElement {
    Space,
    Line(LineMode),
    ExpandParent,
    StaticText {
        text: &'static str,
    },
    DynamicText {
        text: Box<str>,
        source_position: TextSize,
    },
    SyntaxTokenTextSlice {
        source_position: TextSize,
        slice: SyntaxTokenText,
    },
    LineSuffixBoundary,
    Interned(Interned),
    BestFitting(BestFitting),
    Tag(Tag),
}

Language agnostic IR for formatting source code.

Use the helper functions like crate::builders::space, crate::builders::soft_line_break etc. defined in this file to create elements.

Variants

Space

A space token, see crate::builders::space for documentation.

Line(LineMode)

A new line, see crate::builders::soft_line_break, crate::builders::hard_line_break, and crate::builders::soft_line_break_or_space for documentation.

ExpandParent

Forces the parent group to print in expanded mode.

StaticText

Fields

text: &'static str

Token constructed by the formatter from a static string

DynamicText

Fields

text: Box<str>

There’s no need for the text to be mutable, using Box<str> safes 8 bytes over String.

source_position: TextSize

The start position of the dynamic token in the unformatted source code

Token constructed from the input source as a dynamic string with its start position in the input document.

SyntaxTokenTextSlice

Fields

source_position: TextSize

The start position of the token in the unformatted source code

slice: SyntaxTokenText

The token text

A token for a text that is taken as is from the source code (input text and formatted representation are identical). Implementing by taking a slice from a SyntaxToken to avoid allocating a new string.

LineSuffixBoundary

Prevents that line suffixes move past this boundary. Forces the printer to print any pending line suffixes, potentially by inserting a hard line break.

Interned(Interned)

An interned format element. Useful when the same content must be emitted multiple times to avoid deep cloning the IR when using the best_fitting! macro or if_group_fits_on_line and if_group_breaks.

BestFitting(BestFitting)

A list of different variants representing the same content. The printer picks the best fitting content. Line breaks inside of a best fitting don’t propagate to parent groups.

Tag(Tag)

A Tag that marks the start/end of some content to which some special formatting is applied.

Implementations

impl FormatElement

pub const fn is_tag(&self) -> bool

Returns true if self is a FormatElement::Tag

pub const fn is_start_tag(&self) -> bool

Returns true if self is a FormatElement::Tag and Tag::is_start is true.

pub const fn is_end_tag(&self) -> bool

Returns true if self is a FormatElement::Tag and Tag::is_end is true.

pub const fn is_text(&self) -> bool

pub const fn is_space(&self) -> bool