Trait rome_formatter::format_element::FormatElements

pub trait FormatElements {
    fn will_break(&self) -> bool;
    fn has_label(&self, label: LabelId) -> bool;
    fn start_tag(&self, kind: TagKind) -> Option<&Tag>;
    fn end_tag(&self, kind: TagKind) -> Option<&Tag>;
}

Required Methods

fn will_break(&self) -> bool

Returns true if this FormatElement is guaranteed to break across multiple lines by the printer. This is the case if this format element recursively contains a:

crate::builders::empty_line or crate::builders::hard_line_break

A token containing ‘\n’

Use this with caution, this is only a heuristic and the printer may print the element over multiple lines if this element is part of a group and the group doesn’t fit on a single line.

fn has_label(&self, label: LabelId) -> bool

Returns true if the element has the given label.

fn start_tag(&self, kind: TagKind) -> Option<&Tag>

Returns the start tag of kind if:

the last element is an end tag of kind.

there’s a matching start tag in this document (may not be true if this slice is an interned element and the start is in the document storing the interned element).

fn end_tag(&self, kind: TagKind) -> Option<&Tag>

Returns the end tag if:

the last element is an end tag of kind