Struct rome_formatter::formatter::Formatter

pub struct Formatter<'buf, Context> { /* private fields */ }

Handles the formatting of a CST and stores the context how the CST should be formatted (user preferences). The formatter is passed to the Format implementation of every node in the CST so that they can use it to format their children.

Implementations

impl<'buf, Context> Formatter<'buf, Context>

pub fn new(buffer: &'buf mut (dyn Buffer<Context = Context> + 'buf)) -> Self

Creates a new context that uses the given formatter context

pub fn options(&self) -> &Context::Options where
Context: FormatContext,

Returns the format options

pub fn context(&self) -> &Context

Returns the Context specifying how to format the current CST

pub fn context_mut(&mut self) -> &mut Context

Returns a mutable reference to the context.

pub fn group_id(&self, debug_name: &'static str) -> GroupId

Creates a new group id that is unique to this document. The passed debug name is used in the std::fmt::Debug of the document if this is a debug build. The name is unused for production builds and has no meaning on the equality of two group ids.

pub fn join<'a>(&'a mut self) -> JoinBuilder<'a, 'buf, (), Context>

Joins multiple Format together without any separator

Examples

use rome_formatter::format;
use rome_formatter::prelude::*;

let formatted = format!(SimpleFormatContext::default(), [format_with(|f| {
    f.join()
        .entry(&text("a"))
        .entry(&space())
        .entry(&text("+"))
        .entry(&space())
        .entry(&text("b"))
        .finish()
})])?;

assert_eq!(
    "a + b",
    formatted.print()?.as_code()
);

pub fn join_with<'a, Joiner>(
&'a mut self,
joiner: Joiner
) -> JoinBuilder<'a, 'buf, Joiner, Context> where
Joiner: Format<Context>,

Joins the objects by placing the specified separator between every two items.

Examples

Joining different tokens by separating them with a comma and a space.

use rome_formatter::{format, format_args};
use rome_formatter::prelude::*;

let formatted = format!(SimpleFormatContext::default(), [format_with(|f| {
    f.join_with(&format_args!(text(","), space()))
        .entry(&text("1"))
        .entry(&text("2"))
        .entry(&text("3"))
        .entry(&text("4"))
        .finish()
})])?;

assert_eq!(
    "1, 2, 3, 4",
    formatted.print()?.as_code()
);

pub fn join_nodes_with_soft_line<'a>(
&'a mut self
) -> JoinNodesBuilder<'a, 'buf, Line, Context>

Specialized version of crate::Formatter::join_with for joining SyntaxNodes separated by a space, soft line break or empty line depending on the input file.

This functions inspects the input source and separates consecutive elements with either a crate::builders::soft_line_break_or_space or crate::builders::empty_line depending on how many line breaks were separating the elements in the original file.

pub fn join_nodes_with_hardline<'a>(
&'a mut self
) -> JoinNodesBuilder<'a, 'buf, Line, Context>

Specialized version of crate::Formatter::join_with for joining SyntaxNodes separated by one or more line breaks depending on the input file.

This functions inspects the input source and separates consecutive elements with either a crate::builders::hard_line_break or crate::builders::empty_line depending on how many line breaks were separating the elements in the original file.

pub fn fill<'a>(&'a mut self) -> FillBuilder<'a, 'buf, Context>

Concatenates a list of crate::Format objects with spaces and line breaks to fit them on as few lines as possible. Each element introduces a conceptual group. The printer first tries to print the item in flat mode but then prints it in expanded mode if it doesn’t fit.

Examples

use rome_formatter::prelude::*;
use rome_formatter::{format, format_args};

let formatted = format!(SimpleFormatContext::default(), [format_with(|f| {
    f.fill()
        .entry(&soft_line_break_or_space(), &text("aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"))
        .entry(&soft_line_break_or_space(), &text("bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"))
        .entry(&soft_line_break_or_space(), &text("cccccccccccccccccccccccccccccc"))
        .entry(&soft_line_break_or_space(), &text("dddddddddddddddddddddddddddddd"))
        .finish()
})])?;

assert_eq!(
    "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\ncccccccccccccccccccccccccccccc dddddddddddddddddddddddddddddd",
    formatted.print()?.as_code()
);

use rome_formatter::prelude::*;
use rome_formatter::{format, format_args};

let entries = vec![
    text("<b>Important: </b>"),
    text("Please do not commit memory bugs such as segfaults, buffer overflows, etc. otherwise you "),
    text("<em>will</em>"),
    text(" be reprimanded")
];

let formatted = format!(SimpleFormatContext::default(), [format_with(|f| {
    f.fill().entries(&soft_line_break(), entries.iter()).finish()
})])?;

assert_eq!(
    &std::format!("<b>Important: </b>\nPlease do not commit memory bugs such as segfaults, buffer overflows, etc. otherwise you \n<em>will</em> be reprimanded"),
    formatted.print()?.as_code()
);

pub fn intern(
&mut self,
content: &dyn Format<Context>
) -> FormatResult<Option<FormatElement>>

Formats content into an interned element without writing it to the formatter’s buffer.

pub fn intern_vec(
&mut self,
elements: Vec<FormatElement>
) -> Option<FormatElement>

impl<Context> Formatter<'_, Context> where
Context: FormatContext,

pub fn state_snapshot(&self) -> FormatterSnapshot

Take a snapshot of the state of the formatter

pub fn restore_state_snapshot(&mut self, snapshot: FormatterSnapshot)

Restore the state of the formatter to a previous snapshot

impl<Context> Formatter<'_, Context> where
Context: CstFormatContext,

pub fn comments(&self) -> &Comments<Context::Language>

Returns the comments from the context.