Macro rome_formatter::best_fitting

macro_rules! best_fitting {
    ($least_expanded:expr, $($tail:expr),+ $(,)?) => { ... };
}

Provides multiple different alternatives and the printer picks the first one that fits. Use this as last resort because it requires that the printer must try all variants in the worst case. The passed variants must be in the following order:

First: The variant that takes up most space horizontally

Last: The variant that takes up the least space horizontally by splitting the content over multiple lines.

Examples

use rome_formatter::{Formatted, LineWidth, format, format_args};
use rome_formatter::prelude::*;

let formatted = format!(
    SimpleFormatContext::default(),
    [
        token("aVeryLongIdentifier"),
        best_fitting!(
            // Everything fits on a single line
            format_args!(
                token("("),
                group_elements(&format_args![
                    token("["),
                        soft_block_indent(&format_args![
                        token("1,"),
                        soft_line_break_or_space(),
                        token("2,"),
                        soft_line_break_or_space(),
                        token("3"),
                    ]),
                    token("]")
                ]),
                token(")")
            ),

            // Breaks after `[`, but prints all elements on a single line
            format_args!(
                token("("),
                token("["),
                block_indent(&token("1, 2, 3")),
                token("]"),
                token(")"),
            ),

            // Breaks after `[` and prints each element on a single line
            format_args!(
                token("("),
                block_indent(&format_args![
                    token("["),
                    block_indent(&format_args![
                        token("1,"),
                        hard_line_break(),
                        token("2,"),
                        hard_line_break(),
                        token("3"),
                    ]),
                    token("]"),
                ]),
                token(")")
            )
        )
    ]
).unwrap();

let elements = formatted.into_format_element();

// Takes the first variant if everything fits on a single line
assert_eq!(
    "aVeryLongIdentifier([1, 2, 3])",
    Formatted::new(elements.clone(), SimpleFormatContext::default())
        .print()
        .as_code()
);

// It takes the second if the first variant doesn't fit on a single line. The second variant
// has some additional line breaks to make sure inner groups don't break
assert_eq!(
    "aVeryLongIdentifier([\n\t1, 2, 3\n])",
    Formatted::new(elements.clone(), SimpleFormatContext { line_width: 21.try_into().unwrap(), ..SimpleFormatContext::default() })
        .print()
        .as_code()
);

// Prints the last option as last resort
assert_eq!(
    "aVeryLongIdentifier(\n\t[\n\t\t1,\n\t\t2,\n\t\t3\n\t]\n)",
    Formatted::new(elements.clone(), SimpleFormatContext { line_width: 20.try_into().unwrap(), ..SimpleFormatContext::default() })
        .print()
        .as_code()
);

Complexity

Be mindful of using this IR element as it has a considerable performance penalty:

There are multiple representation for the same content. This results in increased memory usage and traversal time in the printer.

The worst case complexity is that the printer tires each variant. This can result in quadratic complexity if used in nested structures.

Behavior

This IR is similar to Prettier’s conditionalGroup. It provides similar functionality but differs in that Prettier automatically wraps each variant in a Group. Rome doesn’t do so. You can wrap the variant content in a group if you want to use soft line breaks. Unlike in Prettier, the printer will try to fit only the first variant in Flat mode, then it will try to fit the rest of the variants in Expanded mode.

A variant that is measured in Expanded mode will be considered to fit if it can be printed without overflowing the current line with all of its inner groups expanded. Those inner groups could still end up being printed in flat mode if they fit on the line while printing. But there is currently no way to enforce that a specific group inside a variant must be flat when measuring if that variant fits.