Function rome_formatter::prelude::align

pub fn align<Content, Context>(
    count: u8,
    content: &Content
) -> Align<'_, Context> where
    Content: Format<Context>, 

Aligns its content by indenting the content by count spaces.

align is a variant of [indent] that indents its content by a specified number of spaces rather than using the configured indent character (tab or a specified number of spaces).

You should use align when you want to indent a content by a specific number of spaces. Using indent is preferred in all other situations as it respects the users preferred indent character.

Examples

Tab indention

use std::num::NonZeroU8;
use rome_formatter::{format, format_args};
use rome_formatter::prelude::*;

let block = format!(SimpleFormatContext::default(), [
    text("a"),
    hard_line_break(),
    text("?"),
    space(),
    align(2, &format_args![
        text("function () {"),
        hard_line_break(),
        text("}"),
    ]),
    hard_line_break(),
    text(":"),
    space(),
    align(2, &format_args![
        text("function () {"),
        block_indent(&text("console.log('test');")),
        text("}"),
    ]),
    text(";")
])?;

assert_eq!(
    "a\n? function () {\n  }\n: function () {\n\t\tconsole.log('test');\n  };",
    block.print()?.as_code()
);

You can see that:

the printer indents the function’s } by two spaces because it is inside of an align.

the block console.log gets indented by two tabs. This is because align increases the indention level by one (same as indent) if you nest an indent inside an align. Meaning that, align > ... > indent results in the same indention as indent > ... > indent.

Spaces indention

use std::num::NonZeroU8;
use rome_formatter::{format, format_args, IndentStyle, SimpleFormatOptions};
use rome_formatter::prelude::*;

let context = SimpleFormatContext::new(SimpleFormatOptions {
    indent_style: IndentStyle::Space(4),
    ..SimpleFormatOptions::default()
});

let block = format!(context, [
    text("a"),
    hard_line_break(),
    text("?"),
    space(),
    align(2, &format_args![
        text("function () {"),
        hard_line_break(),
        text("}"),
    ]),
    hard_line_break(),
    text(":"),
    space(),
    align(2, &format_args![
        text("function () {"),
        block_indent(&text("console.log('test');")),
        text("}"),
    ]),
    text(";")
])?;

assert_eq!(
    "a\n? function () {\n  }\n: function () {\n      console.log('test');\n  };",
    block.print()?.as_code()
);

The printing of align differs if using spaces as indention sequence and it contains an indent. You can see the difference when comparing the indention of the console.log(...) expression to the previous example:

tab indention: Printer indents the expression with two tabs because the align increases the indention level.

space indention: Printer indents the expression by 4 spaces (one indention level) and 2 spaces for the align.