Function rome_formatter::prelude::expand_parent

pub const fn expand_parent() -> ExpandParent

IR element that forces the parent group to print in expanded mode.

Has no effect if used outside of a group or element that introduce implicit groups (fill element).

Examples

use rome_formatter::{format, format_args, LineWidth};
use rome_formatter::prelude::*;

let elements = format!(SimpleFormatContext::default(), [
    group(&format_args![
        text("["),
        soft_block_indent(&format_args![
            text("'Good morning! How are you today?',"),
            soft_line_break_or_space(),
            text("2,"),
            expand_parent(), // Forces the parent to expand
            soft_line_break_or_space(),
            text("3"),
        ]),
        text("]"),
    ])
])?;

assert_eq!(
    "[\n\t'Good morning! How are you today?',\n\t2,\n\t3\n]",
    elements.print()?.as_code()
);

Prettier

Equivalent to Prettier’s break_parent IR element