Function rome_formatter::prelude::format_verbatim_node

pub fn format_verbatim_node<L: Language>(
    node: &SyntaxNode<L>
) -> FormatVerbatimNode<'_, L>

“Formats” a node according to its original formatting in the source text. Being able to format a node “as is” is useful if a node contains syntax errors. Formatting a node with syntax errors has the risk that Rome misinterprets the structure of the code and formatting it could “mess up” the developers, yet incomplete, work or accidentally introduce new syntax errors.

You may be inclined to call node.text directly. However, using text doesn’t track the nodes nor its children source mapping information, resulting in incorrect source maps for this subtree.

These nodes and tokens get tracked as VerbatimKind::Verbatim, useful to understand if these nodes still need to have their own implementation.