Function rome_formatter::prelude::group

pub fn group<Context>(content: &impl Format<Context>) -> Group<'_, Context>

Creates a logical Group around the content that should either consistently be printed on a single line or broken across multiple lines.

The printer will try to print the content of the Group on a single line, ignoring all soft line breaks and emitting spaces for soft line breaks or spaces. The printer tracks back if it isn’t successful either because it encountered a hard line break, or because printing the Group on a single line exceeds the configured line width, and thus it must print all its content on multiple lines, emitting line breaks for all line break kinds.

Examples

Group that fits on a single line

use rome_formatter::{format, format_args};
use rome_formatter::prelude::*;

let elements = format!(SimpleFormatContext::default(), [
    group(&format_args![
        text("["),
        soft_block_indent(&format_args![
            text("1,"),
            soft_line_break_or_space(),
            text("2,"),
            soft_line_break_or_space(),
            text("3"),
        ]),
        text("]"),
    ])
])?;

assert_eq!(
    "[1, 2, 3]",
    elements.print()?.as_code()
);

The printer breaks the Group over multiple lines if its content doesn’t fit on a single line

use rome_formatter::{format, format_args, LineWidth, SimpleFormatOptions};
use rome_formatter::prelude::*;

let context = SimpleFormatContext::new(SimpleFormatOptions {
    line_width: LineWidth::try_from(20).unwrap(),
    ..SimpleFormatOptions::default()
});

let elements = format!(context, [
    group(&format_args![
        text("["),
        soft_block_indent(&format_args![
            text("'Good morning! How are you today?',"),
            soft_line_break_or_space(),
            text("2,"),
            soft_line_break_or_space(),
            text("3"),
        ]),
        text("]"),
    ])
])?;

assert_eq!(
    "[\n\t'Good morning! How are you today?',\n\t2,\n\t3\n]",
    elements.print()?.as_code()
);