Function rome_formatter::prelude::if_group_breaks

pub fn if_group_breaks<Content, Context>(
    content: &Content
) -> IfGroupBreaks<'_, Context> where
    Content: Format<Context>, 

Adds a conditional content that is emitted only if it isn’t inside an enclosing Group that is printed on a single line. The element allows, for example, to insert a trailing comma after the last array element only if the array doesn’t fit on a single line.

The element has no special meaning if used outside of a Group. In that case, the content is always emitted.

If you’re looking for a way to only print something if the Group fits on a single line see self::if_group_fits_on_line.

Examples

Omits the trailing comma for the last array element if the Group fits on a single line

use rome_formatter::{format, format_args};
use rome_formatter::prelude::*;

let elements = format!(SimpleFormatContext::default(), [
    group(&format_args![
        text("["),
        soft_block_indent(&format_args![
            text("1,"),
            soft_line_break_or_space(),
            text("2,"),
            soft_line_break_or_space(),
            text("3"),
            if_group_breaks(&text(","))
        ]),
        text("]"),
    ])
])?;

assert_eq!(
    "[1, 2, 3]",
    elements.print()?.as_code()
);

Prints the trailing comma for the last array element if the Group doesn’t fit on a single line

use rome_formatter::{format_args, format, LineWidth, SimpleFormatOptions};
use rome_formatter::prelude::*;
use rome_formatter::printer::PrintWidth;

fn main() -> FormatResult<()> {
let context = SimpleFormatContext::new(SimpleFormatOptions {
    line_width: LineWidth::try_from(20).unwrap(),
    ..SimpleFormatOptions::default()
});

let elements = format!(context, [
    group(&format_args![
        text("["),
        soft_block_indent(&format_args![
            text("'A somewhat longer string to force a line break',"),
            soft_line_break_or_space(),
            text("2,"),
            soft_line_break_or_space(),
            text("3"),
            if_group_breaks(&text(","))
        ]),
        text("]"),
    ])
])?;

assert_eq!(
    "[\n\t'A somewhat longer string to force a line break',\n\t2,\n\t3,\n]",
    elements.print()?.as_code()
);