Function rome_formatter::prelude::if_group_fits_on_line

pub fn if_group_fits_on_line<Content, Context>(
    flat_content: &Content
) -> IfGroupBreaks<'_, Context>where
    Content: Format<Context>,

Adds a conditional content specific for Groups that fit on a single line. The content isn’t emitted for Groups spanning multiple lines.

See if_group_breaks if you’re looking for a way to print content only for groups spanning multiple lines.

Examples

Adds the trailing comma for the last array element if the Group fits on a single line

use rome_formatter::{format, format_args};
use rome_formatter::prelude::*;

let formatted = format!(SimpleFormatContext::default(), [
    group(&format_args![
        text("["),
        soft_block_indent(&format_args![
            text("1,"),
            soft_line_break_or_space(),
            text("2,"),
            soft_line_break_or_space(),
            text("3"),
            if_group_fits_on_line(&text(","))
        ]),
        text("]"),
    ])
])?;

assert_eq!(
    "[1, 2, 3,]",
    formatted.print()?.as_code()
);

Omits the trailing comma for the last array element if the Group doesn’t fit on a single line

use rome_formatter::{format, format_args, LineWidth, SimpleFormatOptions};
use rome_formatter::prelude::*;

let context = SimpleFormatContext::new(SimpleFormatOptions {
    line_width: LineWidth::try_from(20).unwrap(),
    ..SimpleFormatOptions::default()
});

let formatted = format!(context, [
    group(&format_args![
        text("["),
        soft_block_indent(&format_args![
            text("'A somewhat longer string to force a line break',"),
            soft_line_break_or_space(),
            text("2,"),
            soft_line_break_or_space(),
            text("3"),
            if_group_fits_on_line(&text(","))
        ]),
        text("]"),
    ])
])?;

assert_eq!(
    "[\n\t'A somewhat longer string to force a line break',\n\t2,\n\t3\n]",
    formatted.print()?.as_code()
);