Function rome_formatter::prelude::labelled

pub fn labelled<Content, Context>(
    label_id: LabelId,
    content: &Content
) -> FormatLabelled<'_, Context>where
    Content: Format<Context>,

Marks some content with a label.

This does not directly influence how this content will be printed, but some parts of the formatter may inspect the labelled element using FormatElements::has_label.

Examples

use rome_formatter::prelude::*;
use rome_formatter::{format, write, LineWidth};

enum SomeLabelId {}

let formatted = format!(
    SimpleFormatContext::default(),
    [format_with(|f| {
        let mut recording = f.start_recording();
        write!(recording, [
            labelled(
                LabelId::of::<SomeLabelId>(),
                &text("'I have a label'")
            )
        ])?;

        let recorded = recording.stop();

        let is_labelled = recorded.first().map_or(false, |element| element.has_label(LabelId::of::<SomeLabelId>()));

        if is_labelled {
            write!(f, [text(" has label SomeLabelId")])
        } else {
            write!(f, [text(" doesn't have label SomeLabelId")])
        }
    })]
)?;

assert_eq!("'I have a label' has label SomeLabelId", formatted.print()?.as_code());

Alternatives

Use Memoized.inspect(f)?.has_label(LabelId::of::<SomeLabelId>() if you need to know if some content breaks that should only be written later.