Function rome_formatter::prelude::soft_line_break

pub const fn soft_line_break() -> Line

A line break that only gets printed if the enclosing Group doesn’t fit on a single line. It’s omitted if the enclosing Group fits on a single line. A soft line break is identical to a hard line break when not enclosed inside of a Group.

Examples

Soft line breaks are omitted if the enclosing Group fits on a single line

use rome_formatter::{format, format_args};
use rome_formatter::prelude::*;

let elements = format!(SimpleFormatContext::default(), [
    group(&format_args![text("a,"), soft_line_break(), text("b")])
])?;

assert_eq!(
    "a,b",
    elements.print()?.as_code()
);

See soft_line_break_or_space if you want to insert a space between the elements if the enclosing Group fits on a single line.

Soft line breaks are emitted if the enclosing Group doesn’t fit on a single line

use rome_formatter::{format, format_args, LineWidth, SimpleFormatOptions};
use rome_formatter::prelude::*;

let context = SimpleFormatContext::new(SimpleFormatOptions {
    line_width: LineWidth::try_from(10).unwrap(),
    ..SimpleFormatOptions::default()
});

let elements = format!(context, [
    group(&format_args![
        text("a long word,"),
        soft_line_break(),
        text("so that the group doesn't fit on a single line"),
    ])
])?;

assert_eq!(
    "a long word,\nso that the group doesn't fit on a single line",
    elements.print()?.as_code()
);