Module rome_formatter::token

Structs

FormatInserted

Formats a token that has been inserted by the formatter and isn’t present in the source text. Takes care of correctly handling spacing to the previous token’s trailing trivia.

FormatInsertedCloseParen

Inserts a closing parentheses before another token and moves that tokens trailing trivia after the closing parentheses.

FormatInsertedOpenParen

Inserts a open parentheses before the specified token and ensures that any leading trivia of the token is formatted before the inserted parentheses.

FormatLeadingTrivia

FormatOnlyIfBreaks

Formats a token with its leading and trailing trivia that only gets printed if its enclosing group does break but otherwise gets omitted from the formatted output.

FormatRemoved

Formats the trivia of a token that is present in the source text but should be omitted in the formatted output.

FormatReplaced

Formats a token’s leading and trailing trivia but uses the provided content instead of the token in the formatted output.

FormatTrailingTrivia

FormatTrimmedToken

Enums

TriviaPrintMode

Determines if the whitespace separating comment trivia from their associated tokens should be printed or trimmed

Functions

format_leading_trivia

Formats the leading trivia (comments, skipped token trivia) of a token

format_only_if_breaks

Formats the given token only if the group does break and otherwise retains the token’s trivia.

format_removed

Formats the leading and trailing trivia of a removed token.

format_replaced

Print out a token from the original source with a different content.

format_trailing_trivia

Formats the trailing trivia (comments) of a token

format_trimmed_token

! Provides builders for working with tokens and the tokens trivia Formats a token without its leading or trailing trivia