Function rome_formatter::token::format_trimmed_token

pub const fn format_trimmed_token<L: Language>(
    token: &SyntaxToken<L>
) -> FormatTrimmedToken<'_, L>

! Provides builders for working with tokens and the tokens trivia Formats a token without its leading or trailing trivia

Warning

It’s your responsibility to format leading or trailing comments and skipped trivia.