Trait rome_formatter::BufferExtensions

pub trait BufferExtensions: Buffer + Sized {
    fn inspect<F>(&mut self, inspector: F) -> Inspect<'_, Self::Context, F>
    where
        F: FnMut(&FormatElement),
    { ... }
    fn write_elements<I>(&mut self, elements: I) -> FormatResult<()>
    where
        I: IntoIterator<Item = FormatElement>,
    { ... }
    fn inspect_will_break(&mut self) -> WillBreakBuffer<'_, Self::Context> { ... }
    fn inspect_is_labelled<T: ?Sized + 'static>(
        &mut self
    ) -> IsLabelledBuffer<'_, Self::Context> { ... }
    fn inspect_null(&mut self) -> NullBuffer<'_, Self::Context> { ... }
    fn label_of<T: ?Sized + 'static>(&self) -> LabelId { ... }
}

Provided Methods

fn inspect<F>(&mut self, inspector: F) -> Inspect<'_, Self::Context, F> where
F: FnMut(&FormatElement),

Returns a new buffer that calls the passed inspector for every element that gets written to the output

fn write_elements<I>(&mut self, elements: I) -> FormatResult<()> where
I: IntoIterator<Item = FormatElement>,

Writes a sequence of elements into this buffer.

fn inspect_will_break(&mut self) -> WillBreakBuffer<'_, Self::Context>

It emits a custom buffer called WillBreakBuffer, which tracks it he last element written in the main buffer breaks, it does so by checking if their IR emits an element that breaks.

This functionality can be used only one element and only after the element is written in the buffer.

Examples

use rome_formatter::{format, format_args, write, LineWidth};
use rome_formatter::prelude::*;

let context = SimpleFormatContext {
    line_width: LineWidth::try_from(20).unwrap(),
    ..SimpleFormatContext::default()
};

let formatted = format!(context, [format_with(|f| {
     
    let element = format_with(|f| {
        write!(f, [
            token("hello"),
            hard_line_break(),
            token("world!")
        ])
    });
    let mut buffer = f.inspect_will_break();
    write!(buffer, [element])?;
    let does_element_break = buffer.will_break();

    if does_element_break {
        write!(f, [hard_line_break(), token("break")])
    } else {
        write!(f, [token("did not break")])
    }   
    
})]).unwrap();

assert_eq!(
    "hello\nworld!\nbreak",
    formatted.print().as_code()
);

fn inspect_is_labelled<T: ?Sized + 'static>(
&mut self
) -> IsLabelledBuffer<'_, Self::Context>

It emits a custom buffer called IsLabelledBuffer, which tracks labelled elements written in the main buffer, it does so by checking if element is label with expected label_id.

This functionality can be used only on one element and only after the element is written in the buffer.

Examples

use rome_formatter::prelude::*;
use rome_formatter::{format, write, LineWidth};

enum SomeLabelId {}

let context = SimpleFormatContext {
    line_width: LineWidth::try_from(20).unwrap(),
    ..SimpleFormatContext::default()
};

let formatted = format!(
    context,
    [format_with(|f| {
        let element = format_with(|f| {
            let label_id = f.label_of::<SomeLabelId>();

            write!(f, [labelled(label_id, &token("labelled")),])
        })
        .memoized();

        let is_labelled = {
            let mut null_buffer = f.inspect_null();
            let mut buffer = null_buffer.inspect_is_labelled::<SomeLabelId>();
            write!(buffer, [element])?;
            buffer.is_labelled()
        };

        if is_labelled {
            write!(f, [token("This is "), &element])
        } else {
            write!(f, [token("This is not "), &element])
        }
    })]
)
.unwrap();

assert_eq!("This is labelled", formatted.print().as_code());

fn inspect_null(&mut self) -> NullBuffer<'_, Self::Context>

It creates a buffer where all the elements are ignored, so the elements are not written anywhere at all.

This can be useful when formatters are not yet written inside the main buffer and the consumer needs to inspect them, to decide the formatting layout in advance.

Examples

The following example shows how to use it with the will_break functionality

use rome_formatter::{format, format_args, write, LineWidth};
use rome_formatter::prelude::*;

let context = SimpleFormatContext {
    line_width: LineWidth::try_from(20).unwrap(),
    ..SimpleFormatContext::default()
};

let formatted = format!(context, [format_with(|f| {

    let element = format_with(|f| {
        write!(f, [
            token("hello"),
            hard_line_break(),
            token("world!")
        ])
    }).memoized();

    let will_break =  {
        let mut null_buffer =  f.inspect_null();
        let mut buffer = null_buffer.inspect_will_break();
        write!(buffer, [element])?;
        buffer.will_break()
    };

    if will_break {
        write!(f, [token("break"), hard_line_break(), &element])
    } else {
        write!(f, [token("did not break")])
    }   
    
})]).unwrap();

assert_eq!(
    "break\nhello\nworld!",
    formatted.print().as_code()
);

fn label_of<T: ?Sized + 'static>(&self) -> LabelId