Trait rome_formatter::CommentStyle

pub trait CommentStyle<L: Language> {
    fn is_suppression(&self, text: &str) -> bool;
    fn get_comment_kind(
        &self, 
        comment: &SyntaxTriviaPieceComments<L>
    ) -> CommentKind;
    fn is_group_start_token(&self, kind: L::Kind) -> bool;
    fn is_group_end_token(&self, kind: L::Kind) -> bool;
}

Defines how to format comments for a specific Language.

Required Methods

fn is_suppression(&self, text: &str) -> bool

Returns true if a comment with the given text is a rome-ignore format: suppression comment.

fn get_comment_kind(
&self,
comment: &SyntaxTriviaPieceComments<L>
) -> CommentKind

Returns the (kind)CommentKind of the comment

fn is_group_start_token(&self, kind: L::Kind) -> bool

Returns true if a token with the passed kind marks the start of a group. Common group tokens are:

left parentheses: (, [, {

fn is_group_end_token(&self, kind: L::Kind) -> bool

Returns true if a token with the passed kind marks the end of a group. Common group end tokens are:

right parentheses: ), ], }

end of statement token: ;

element separator: , or ..

end of file token: EOF