Trait rome_formatter::CstFormatContext

pub trait CstFormatContext: FormatContext {
    type Language: Language;
    type Style: CommentStyle<Self::Language>;

    fn comment_style(&self) -> Self::Style;
    fn comments(&self) -> Rc<Comments<Self::Language>>;
    fn with_comments(self, comments: Rc<Comments<Self::Language>>) -> Self;
}

The CstFormatContext is an extension of the CST unaware FormatContext and must be implemented by every language.

The context customizes the comments formatting and stores the comments of the CST.

Required Associated Types

type Language: Language

type Style: CommentStyle<Self::Language>

Required Methods

fn comment_style(&self) -> Self::Style

Customizes how comments are formatted

fn comments(&self) -> Rc<Comments<Self::Language>>

Returns a ref counted Comments.

The use of a Rc is necessary to achieve that Comments has a lifetime that is independent of the crate::Formatter. Having independent lifetimes is necessary to support the use case where a (formattable object)Format iterates over all comments and writes them into the crate::Formatter (mutably borrowing the crate::Formatter and in turn this context).

for leading in f.context().comments().leading_comments(node) {
    ^
    |- Borrows comments
  write!(f, [comment(leading.piece.text())])?;
         ^
         |- Mutably borrows the formatter, state, context (and comments, if they aren't wrapped by a Rc)
}

fn with_comments(self, comments: Rc<Comments<Self::Language>>) -> Self

Consumes self and returns a new context with the provided extracted (comments)Comments.