Trait rome_formatter::FormatContext

pub trait FormatContext {
    type Snapshot;

    fn indent_style(&self) -> IndentStyle;
    fn line_width(&self) -> LineWidth;
    fn as_print_options(&self) -> PrinterOptions;
    fn snapshot(&self) -> Self::Snapshot;
    fn restore_snapshot(&mut self, snapshot: Self::Snapshot);
}

Context configuring how an object gets formatted.

Defines the common formatting options. Implementations can define additional options that are specific to formatting a specific object.

Required Associated Types

type Snapshot

The type storing a snapshot of the context’s (mutable) state.

Required Methods

fn indent_style(&self) -> IndentStyle

The indent style.

fn line_width(&self) -> LineWidth

What’s the max width of a line. Defaults to 80.

fn as_print_options(&self) -> PrinterOptions

Derives the print options from the these format options

fn snapshot(&self) -> Self::Snapshot

Snapshots the state of the context that can be restored with restore_snapshot.

fn restore_snapshot(&mut self, snapshot: Self::Snapshot)

Restores the contexts state to the state when the snapshot was taken.