Trait rome_formatter::FormatLanguage

pub trait FormatLanguage {
    type SyntaxLanguage: Language;
    type Context: CstFormatContext<Language = Self::SyntaxLanguage>;
    type CommentStyle: CommentStyle<Language = Self::SyntaxLanguage>;
    type FormatRule: FormatRule<SyntaxNode<Self::SyntaxLanguage>, Context = Self::Context> + Default;

    fn options(&self) -> &<Self::Context as FormatContext>::Options;
    fn create_context(
        self,
        comments: Comments<Self::SyntaxLanguage>,
        source_map: Option<TransformSourceMap>
    ) -> Self::Context;

    fn transform(
        &self,
        _root: &SyntaxNode<Self::SyntaxLanguage>
    ) -> Option<(SyntaxNode<Self::SyntaxLanguage>, TransformSourceMap)> { ... }
    fn is_range_formatting_node(
        &self,
        _node: &SyntaxNode<Self::SyntaxLanguage>
    ) -> bool { ... }
}

Entry point for formatting a SyntaxNode for a specific language.

Required Associated Types

type SyntaxLanguage: Language

type Context: CstFormatContext<Language = Self::SyntaxLanguage>

The type of the formatting context

type CommentStyle: CommentStyle<Language = Self::SyntaxLanguage>

The type specifying how to format comments.

type FormatRule: FormatRule<SyntaxNode<Self::SyntaxLanguage>, Context = Self::Context> + Default

The rule type that can format a SyntaxNode of this language

Required Methods

fn options(&self) -> &<Self::Context as FormatContext>::Options

Returns the formatting options

fn create_context(
self,
comments: Comments<Self::SyntaxLanguage>,
source_map: Option<TransformSourceMap>
) -> Self::Context

Creates the FormatContext with the given source map and comments

Provided Methods

fn transform(
&self,
_root: &SyntaxNode<Self::SyntaxLanguage>
) -> Option<(SyntaxNode<Self::SyntaxLanguage>, TransformSourceMap)>

Performs an optional pre-processing of the tree. This can be useful to remove nodes that otherwise complicate formatting.

Return None if the tree shouldn’t be processed. Return Some with the transformed tree and the source map otherwise.

fn is_range_formatting_node(
&self,
_node: &SyntaxNode<Self::SyntaxLanguage>
) -> bool

This is used to select appropriate “root nodes” for the range formatting process: for instance in JavaScript the function returns true for statement and declaration nodes, to ensure the entire statement gets formatted instead of the smallest sub-expression that fits the range