Trait rome_formatter::FormatOptions

pub trait FormatOptions {
    fn indent_style(&self) -> IndentStyle;
    fn line_width(&self) -> LineWidth;
    fn as_print_options(&self) -> PrinterOptions;
}

Options customizing how the source code should be formatted.

Required Methods

fn indent_style(&self) -> IndentStyle

The indent style.

fn line_width(&self) -> LineWidth

What’s the max width of a line. Defaults to 80.

fn as_print_options(&self) -> PrinterOptions

Derives the print options from the these format options