Trait rome_formatter::FormatWithRule

pub trait FormatWithRule<Context>: Format<Context> {
    type Item;

    fn item(&self) -> &Self::Item;
}

Trait for an object that formats an object with a specified rule.

Gives access to the underlying item.

Useful in situation where a type itself doesn’t implement Format (e.g. because of Rust’s orphan rule) but you want to implement some common formatting logic.

Examples

This can be useful if you want to format a SyntaxNode inside rome_formatter.. SyntaxNode doesn’t implement Format itself but the language specific crate implements AsFormat and IntoFormat for it and the returned Format implement FormatWithRule.

use rome_formatter::prelude::*;
use rome_formatter::{format, Formatted, FormatWithRule};
use rome_rowan::{Language, SyntaxNode};
fn format_node<L: Language, F: FormatWithRule<SimpleFormatContext, Item=SyntaxNode<L>>>(node: F) -> FormatResult<Formatted<SimpleFormatContext>> {
    let formatted = format!(SimpleFormatContext::default(), [node]);
    let syntax = node.item();
    // Do something with syntax
    formatted;
}

Required Associated Types

type Item

Required Methods

fn item(&self) -> &Self::Item

Returns the associated item