Module rome_formatter::trivia

Provides builders for comments and skipped token trivia.

Structs

FormatOnlyIfBreaks

Formats a token with its skipped token trivia that only gets printed if its enclosing group does break but otherwise gets omitted from the formatted output.

FormatRemoved

Formats the trivia of a token that is present in the source text but should be omitted in the formatted output.

FormatReplaced

Formats a token’s skipped token trivia but uses the provided content instead of the token in the formatted output.

FormatSkippedTokenTrivia

Formats the skipped token trivia of token.

FormatTrimmedToken

Enums

DanglingIndentMode

FormatDanglingComments

Formats the dangling trivia of token.

FormatLeadingComments

Formats the leading comments of a node.

FormatTrailingComments

Formats the trailing comments of node

Functions

format_dangling_comments

Formats the dangling comments of node.

format_leading_comments

Formats the leading comments of node

format_only_if_breaks

Formats the given token only if the group does break and otherwise retains the token’s skipped token trivia.

format_removed

Formats the skipped token trivia of a removed token and marks the token as tracked.

format_replaced

Print out a token from the original source with a different content.

format_skipped_token_trivia

Formats the skipped token trivia of token.

format_trailing_comments

Formats the trailing comments of node.

format_trimmed_token

Formats a token without its skipped token trivia